Move stress tests to their own file
